MODE-775 Previously, when there were errors in 2+ Request objects within a single CompositeRequest, an error message summarizing the request exceptions was built, a RequestException was created with this error message, and this exception set on the CompositeRequest.
Now, in this same situation, the Request objects with the errors are accumulated in a list and wrapped in a MultipleRequestFailuresException (a new class) and set as the CompositeRequest's error. This exception's printable message is hopefully more readable, plus since the original Request objects are contained in the exception, it is much more amenable to debugging.

Also removed System.out.println calls within unit tests.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2369 76366958-4244-0410-ad5e-bbfabb93f86b