MODE-954 I don't think the previous fix actually addressed the problem. If the index of the full-text scores in the graph QueryResults is -1, that means the scores were not being included in the results. Right now, they are only included in certain situations (e.g., the 'SCORE' function is used, or there is a full-text search constraint), but apparently not when the 'jcr:score' pseudo-column appears in the SELECT.
The easiest way to fix this is to always include the scores in the query results, even if it is not needed. (The alternative would be to look at the query to see if the 'jcr:score' pseudo-column appears in the SELECT, and since this is not exactly trivial, it's far easier to just always include the score.) Recall that in Lucene, the score is always computed, so including the score in the results basically just adds a single column per selector. See the changes in the 'MODE-954-fix2.patch' attachment.

All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2403 76366958-4244-0410-ad5e-bbfabb93f86b