MODE-959 After some debugging with bcarothers, we figured out a potential solution. In some cases, the root node will have no properties (not even a jcr:primaryType property) in the graph level, and the JCR layer presumes 'mode:root' as the primary type for the root node. So by checking if the root path was included when the propertiesByLocation map was being populated, and then after it was populated by getting information from the graph, the code can ensure that there is a primary type for the root node.
Investigating this also led to the discovery that the primary type and mixin type were being found for the wrong node for ChangeRequest objects that only dealt with property changes. In such requests, ChangeRequest.changedLocation() will return the path of the node that was changed, but the JcrObservationManager was always getting that node's parent. A simple if-check provides the correct behavior, as well as checking during the 'acceptBasedOnNodeTypeName(...)' method.

A new test case was created that verified the problem (before changes) and verified the fix (after the changes).

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2412 76366958-4244-0410-ad5e-bbfabb93f86b