MODE-973 FileSystemConnector behaves strangely on NFS
Applied patch that modifies the FS connector to fallback to a copy/delete strategy if calls to renameTo fail.  The copy/delete approach will work if the renameTo fails due to the source and target of the rename being in different file systems.  If the copy/delete approach fails, an exception is still thrown.

All tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2444 76366958-4244-0410-ad5e-bbfabb93f86b