MODE-970, MODE-975, MODE-934 Changed the behavior of SELECT * so that it now expands to include pseudo-columns. Several test cases needed to be changed, since the columns returned in the result sets have changed. Also added new test cases that verify the behavior, and verify the ability for the JDBC driver to query all tables returned in the DatabaseMetaData.
In the process of running these new tests, an issue was found and fixed in the canonical planner that caused an exception when a column representing an existing property and one of the pseudo-columns had the same name. (In particular, 'nt:propertyDefinition' has a 'jcr:name' property.)

All unit and integration tests now pass with these changes. The Reference Guide was also updated to reflect this new behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2446 76366958-4244-0410-ad5e-bbfabb93f86b