MODE-919 Getting a broken pipe exception when publishing the attached DDL file
Applied patch that adds a new, optional query parameter (mode:includeNode) to POSTs of JCR items.  If this parameter is passed in with a value that equates to true, then only the path of the new item is returned in the response.  If this parameter is not set or is set but does not evalute to true, then the prior behavior of returning a JSON-encoded version of the newly created node is used.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2503 76366958-4244-0410-ad5e-bbfabb93f86b