MODE-986 FileSystemConnector should perform validation on filenamefilter
Applied patch that adds checks to prevent creating files that will be excluded based on the exclusion filter (or the custom properties filter).  Attempts to create a file with an excluded name through node creation, renaming, or copying should now fail.  Additional test cases to verify this behavior were added.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2507 76366958-4244-0410-ad5e-bbfabb93f86b