MODE-983 Transaction isolation level cannot be configured for ModeShape JPA connector
Applied patch that adds an isolationLevel property to the JPA store.  This property can be set with any value from java.sql.Connection.TRANSACTION_*, although not all values are likely to work.  This value gets set on the JPA connection.

The patch adds a test that verifies that the isolation level is set, but the test relies on the SimpleJpaConnection exposing the underlying JDBC connection.  This requires a hard-coded cast to a Hibernate JPA implementation class.  Rather than add this brittleness to the codebase, I have added the test (and the supporting method in SimpleJpaConnection), run it manually to verify that things work, and then commented out the test and the supporting method.

The patch updates the Reference Guide and is ready to commit if there is agreement that the previous limitation is not significant.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2509 76366958-4244-0410-ad5e-bbfabb93f86b