MODE-923 Performance problem, MapTransaction.findNode() called excessively
Applied patch that adds a node cache to each MapTransaction.  The cache stores nodes that have already been read during that transaction, mitigating the effects of repeated attempts to read the same node.  This is based on the idea that Randall proposed in his earlier post on this defect.

Some quick experimentation with this patch showed a 70% or more reduction in the number of node accesses across most of the InMemoryConnectorReadableTest and InMemoryConnectorWritableTest cases.  Because of the way that the map-based connectors share code, I think it is reasonable safe to extrapolate comparable reductions in reads for the Infinispan connector as well.

All tests still pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2515 76366958-4244-0410-ad5e-bbfabb93f86b