MODE-1012
The supplied patch was a good attempt at tracking down the problem and proving that using the correct source for the output would indeed fix the problem (thanks again!). But because of the way the SequencingService can deal with multiple output paths for a single sequencer invocation, and because each of those output paths might be in a different source, a different approach was necessary.

The fix organizes the set of output paths (i.e., RepositoryNodePath objects) for a given sequencer call by source name. (A new 'SequencerCalls' internal class was used to encapsulate the Map<SequencerCall,Map<String,Set<RepositoryNodePath>>>.) Doing this, the sequencer can be executed once for each output source, and the output of the sequencing written to each of the output paths in that source. A new test case was added to verify that the sequenced output is indeed able to be written to a different repository source than the input.

A few other minor changes were made at the same time.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2528 76366958-4244-0410-ad5e-bbfabb93f86b