MODE-1019 Implemented the ability to specify in the configuration file initial content that should appear in the repository upon startup. If that content does not exist (or exists and does not exactly match the content specified in the initial content definition), the repository content is changed so that this content is reflected. Usually this entails creating new nodes or changing the properties of existing nodes.
The following fragment shows an example of what is now possible to specify within the ModeShape configuration:

<mode:repository ...>
   <mode:initialContent workspaces="ws1" applyToNewWorkspaces="true" content="path/to/content.xml"/>
   <mode:initialContent content="path/to/content.xml">
      <workspaces>ws2</workspaces>
      <workspaces>ws3</workspaces>
      <workspaces>ws4</workspaces>
   </mode:initialContent>
</mode:repository>

Multiple "mode:initialContent" elements are allowed, enabling multiple initial content files to be used for various workspaces.

Note that comma-separated workspace names are **not** supported; instead, use the attribute for a single workspace name or separate child elements for each workspace name.

Also, the 'content' value can be an absolute or relative file system path, the path to a classpath resource, or a resolvable URL. (Values are resolved in that order.)

The last 'initialContent' element that has a 'applyToNewWorkspaces' property set to 'true' will win. This signifies that the initial content check should be performed on workspaces that are created programmatically via the JCR API.

And at this time, the format of the XML is the same as that used by the graph-level XML import. There are two reasons:
i) implementing this as a JCR-level function requires a valid session, whereas the lower-level and session-less mechanism to manipulate content is the graph API
ii) strictly speaking, this is not an 'import' operation, but is closer to a 'merge' of some specified content into the repository workspace. That is, the repository might contain some of the content specified in the 'initialContent' definition, and this process will modify that existing content so that it fully matches that specified in the 'initialContent' definition.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2539 76366958-4244-0410-ad5e-bbfabb93f86b