MODE-1023 Changed JcrValue to enforce its 'value' field is correctly set to an object whose class matched that defined by the desired PropertyType value. Previously, the JcrValue just used whatever Object was supplied.
When this code was added, several dozen TCK and unit test started failing. But upon investigation, it was determined that there were two as-of-yet-undiscovered bugs (one in JcrQueryManager and one in AbstractJcrNode) where the wrong PropertyType was supplied to the JcrValue constructor. The correction involves simply using the correct PropertyType.

After these changes were made, all unit and integration tests pass again.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2549 76366958-4244-0410-ad5e-bbfabb93f86b