MODE-1013 The problem was in fact the XML sequencer was producing a 'modexml:document' node named the same as the XML file, and the configuration was generating a node with the same name. Unfortunately, changing the configuration to no longer produce a node with the same name caused the node produced by the XML sequencer to be named incorrectly, and thus this was not an option.
The fix is to change the configuration and correct the bug in the XML sequencer, and I chose to follow the same pattern used in many of the other sequencers, where the top-level node is generated from the sequenced node or the parent node if the sequenced node is named 'jcr:content' (meaning the parent is likely a 'nt:file' node). Naturally, this meant quite a few of the test cases needed to change their expected values.

Also included in this change is a correction/enhancement to the StreamSequencerAdapter, which is the class that is responsible for reading the file content, calling the StreamSequencer implementation (most of our sequencers implement this simpler interface), and then take the output generated by the StreamSequencer and write it into the repository as a single graph batch operation. With this change, the StreamSequencerAdapter now examines the generated output (specifically the names of the top-level node or nodes in the output) and looks for existing nodes of the same name. If there already are existing nodes, the generated nodes have the incorrect SNS indexes, and this fix modifies the generated output to have the correct path with the next unused SNS index. The result is a much more consistent handling of the generated content when there is already existing content with similar names. A few new test cases were added and several existing were changed to verify this new behavior.

All unit and integration tests pass with these changes.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2573 76366958-4244-0410-ad5e-bbfabb93f86b