MODE-1036 The JPA Connector by default uses TRANSACTION_REPEATABLE_READ, which is not supported in Oracle. Of course, you can easily change this for your database and use either TRANSACTION_READ_COMMITTED or TRANSACTION_SERIALIZABLE by setting the "isolationLevel" property to a value of 2 or 4 (respectively) on the appropriate "mode:source" fragment (in the ModeShape configuration).
The proposal is to change the JPA connector to use a default isolation level that means "don't set this; use whatever the datasource's default is" (likely a null value, since the JavaBean method takes an Integer). If we don't, all Oracle users will be required to specify a value, which seems unattractive since most of our customers are likely using Oracle or MySQL.

I've changed the default value of the 'isolationLevel' JavaBean property on the JpaSource class. Now, the default value is null, meaning the JpaSource will **not** explicitly set the 'hibernate.connection.isolation' property, so the JDBC DataSource's own setting or default will be used. However, if this JavaBean property is set to 0, 1, 2, 4 or 8 (corresponding to the integer values of the standard JDBC transaction isolation levels), the connector **will** set the 'hibernate.connection.isolation' property to this value.

All unit and integration tests pass with these changes. (In fact, I don't believe any of our tests explicitly set this property, so they should be testing with the DataSource's default value.)

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2579 76366958-4244-0410-ad5e-bbfabb93f86b