MODE-1035 The NPE is occurring on line 200 of LargeValueEntity:
        if (dialect.toLowerCase().indexOf("mysql") != -1) {

where 'dialect' is passed into the deleteUnused(...) method from the value set on the JpaSource. Looking in Aleksandar's configuration file (attached), the 'mode:dialect' value is not set. Thus the NPE.

Now, the workaround is to explicitly set the dialect. However, as Aleksandar correctly points out, Hibernate is capable of automatically determining the dialect. (Who knew?!) Unfortunately, it's the ModeShape code (in LargeValueEntity) that is using the dialect value to determine the behavior, so this means we need to ask Hibernate for the dialect after it's made the connection, and set it on the JpaSource. (That way, all components that ask for the dialect, including SimpleJpaRepository, will get the correct dialect name.) See the attached MODE-1035-fix.patch for details.

Note that one of the changes was to explicitly throw an exception if the dialect is not set and could not be determined.

As of this change, setting the dialect is no longer required.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2584 76366958-4244-0410-ad5e-bbfabb93f86b