MODE-1038 The configuration file for ModeShape's JBoss AS kit also had this same error, so the options were changed to use the mode:option XML element style.
Also, when the configuration file specified a name of the JAAS application policy that did not exist, the error message in the javax.jcr.LoginException thrown by JcrSession.login(...) methods did not mention the name of the policy. The logic in JcrRepository was changed to make the underlying problem much clearer.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2587 76366958-4244-0410-ad5e-bbfabb93f86b