MODE-1028 Implemented this with a new internal graph 'net-change' observer on the JcrRepository instance. When the observer receives delete or move events, it queries the appropriate workspace for the 'mode:derived' nodes with a 'mode:derivedFrom' value that matches the path of the deleted or moved node(s) (using LIKE, as the deleted/moved nodes might be ancestors of the nodes pointed to by the 'mode:derivedFrom' value) and with a 'mode:derivedAt' timestamp that is on or before the delete/move event's timestamp (this ensures that the 'mode:derived' nodes were not added in a subsequent operation (e.g., the file is delete and then re-added).
Note that 'mode:derivedAt' is a new DATE property that is now added by the StreamSequencerAdapter (the same code that adds the 'mode:derived' mixin.

A new integration test was added to verify this new logic is performed correctly.

Also, a JcrRepository.Option was added to control whether derived content should be removed. By default (when at least one sequencer is enabled), this new feature is enabled, meaning the repository will remove all content that was derived from other content being removed. However, if no sequencers are enabled, the default is to not enable this feature. Of course, the feature can be explicitly enabled or disabled by specifying the desired value for this option.

Also added a new section to the Reference Guide that describes the JcrRepository.Option values.

All unit and integration tests pass with these changes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2629 76366958-4244-0410-ad5e-bbfabb93f86b