MODE-1066, MODE-1071 Added a new background garbage collection facility to the connector API and ModeShape engine
This new facility includes:
- a new CollectGarbageRequest, with a field allowing the connector to specify whether the garbage collection
  operation was completed or whether another pass could/should be made by the requestor
- a default implementation of the RequestProcessor.process(CollectGarbageRequest) that does nothing; existing
  connectors by default implement this behavior and don't need to change
- a new field on RepositorySourceCapabilities that allows the connector to express whether it automatically
  cleans up unused content (collects garbage), or whether it needs to be done with an explicit CollectGarbageRequest
- a new method on RepositoryService (which owns the library of RepositorySource instances) that will find the
  RepositorySource instances that require explicit garbage collection, and upon these (if there are any)
  iteratively make the CollectGarbageRequest; the source will be enqueued again if another collection pass is required
- a new background scheduled executor in ModeShapeEngine (the superclass of JcrEngine) to periodically (and
  in a background thread) have the RepositoryService collect garbage on its sources.
- a new configuration property that defines the time interval for the periodic, background garbage collection
  (the default is every 10 minutes).

Also changed the JPA connector to override the RequestProcessor.process(CollectGarbageRequest) by removing all
unused LargeValueEntity records in all workspaces, and to declare  in its RepositorySourceCapabilities that manual
GC is required. Since this is now done explicitly, removing nodes no longer removes unused LargeValueEntity records.
This should improve the performance of certain operations, and because the LargeValueEntity records (keyed by
their SHA-1) stick around for a longer period of time, they remain available to be reused. For example, if repository
content is removed but replaced with mostly similar content, any large values (strings or binary property values)
that are the same actually would be reused.

Also, the MySQL-specific logic for removing LargeValueEntity (see MODE-691) was changed to use a single native SQL
delete statement. This is much more efficient, and on-par with the HQL delete statement used for other Hibernate
dialects. The native MySQL statement was needed because HQL was inadequate for deleting the unused LargeValueEntity
without resorting to a subquery that used the LargeValueEntity table, which is not supported by MySQL. So instead,
the native MySQL delete statement uses a left outer join between the LargeValueEntity and "usages" table and a
criteria on the result to ensure that the only records returned are those without any "usages" records in the
tuples. This works because a left outer join between tables A and B always contains all records from A, whether
or not there are no corresponding records from B. And, if a criteria is added such that a column from B that
can never be null is actually null, then we know the result will contain only those records from A that
do **not** have a corresponding B record. In our case, this ends up deleting only those LargeValueEntity records
that are not referenced in the "usages" table (i.e., they are not used).

Updated a couple of test cases and added a few more that deal with the new feature. A new "mysql5_local" database
configuration was added to the POM, and this required modification of several test cases that depended on the
database configuration name to find files in test data.

All unit and integration tests pass while using HSQLDB and MySQL5.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2668 76366958-4244-0410-ad5e-bbfabb93f86b