MODE-1077 Corrected the indexing logic to prevent nodes from not appearing in the indexes.
I was able to replicate this faulty behavior with an integration test that was sequencing a file and querying
the results, only after configuring the engine to use 'initial content'. In all our testing, we were able to
show that the content in the store was correct, and that the only problem appeared to be that some nodes were
missing from the indexes.

It turns out that the problem is in the 'modeshape-search-lucene' module's logic for how the ChangeRequest
objects are processed and used to update the indexes. Using 'initial content' changed the requests that were
being generated such that this condition occurred; when no 'initial content' was specified, the ChangeRequest
objects were different enough that this condition was not encountered.

The logic to process the ChangeRequests tries to figure out whether each change request contains all the
information necessary to update the indexes. Some changes requests, like CreateNodeRequest or
DeleteSubgraphRequest or AddPropertyRequest, have all the information necessary to update the indexes. Other
ChangeRequests don't have all the information necessary (e.g., an UpdatePropertiesRequest doesn't have the
old values, which we have to remove from the indexes), so in these cases we crawl the content to retrieve the
information and replace the record for that node. For many of the requests where we don't have sufficient
information, we crawl just that node (e.g., a crawl depth of 1), but in a few cases we determine that it's more
effective just to crawl the entire subgraph and replace all of the index entries for the entire subgraph. When
crawling the entire subgraph, we can safely ignore all those ChangeRequests that have a changed location below
the crawled paths, so we prune these requests.

The problem was that we were pruning the requests that applied below the crawled path
**even when we were just crawling to a depth of 1**. That means we were pruning some requests, and the
indexes would not reflect those nodes. Only if the subgraph containing those nodes were re-crawled would
the nodes reappear, but that would only happen occasionally.

The fix is actually pretty simple. In the LuceneSearchEngine, where we are pruning the change requests that
apply below the crawl location, we simply have to check the depth and only prune if the depth > 1.

Several test cases were added in the attempt to replicate, and these tests now pass. In fact, all unit and
integration tests pass after the changes are made.