MODE-1078 Corrected the RepositoryLockManager.notify() method to use the correct "mode:workspace" property.
Indeed, the "mode:workspaceName" property (e.g., the ModeShapeLexicon.WORKSPACE_NAME constant) is used in
the configuration of the federation definitions, while "mode:workspace" is used within the lock manager
implementation. Therefore, the RepositoryLockManager.notify(...) method should be using
ModeShapeLexicon.WORKSPACE (the same as is already used in other methods of the class.

Also added an assertion to the RepositoryLockManager.getLockManager(...) method to ensure the workspace
name is not null, and added this pre-condition to the existing JavaDoc.

All unit and integration tests pass.