MODE-1073 Corrected the sequencing service's logic of creating intermediary nodes above the sequenced output to be more thread-safe.
It looks like this is a problem with the Sequencing Service and not the DDL sequencer. The problem appears to be
caused by a race condition between multiple sequencers running at once (in separate threads), and of course is
related to how quickly the files are uploaded (published) and how busy the service already is. For example, if
the files are published a bit more slowly, the threads are far less likely to run concurrently, and the problem
will not be seen.

I was able to replicate this with two new test cases that almost always exhibited the problem. (It is a race
condition, after all.)

However, the cause is actually the code that creates the intermediary nodes (the nodes about the sequenced output).
This code looks to see if the nodes already exists, and if so then it doesn't recreated them. If two sequencers
are running concurrently looking for the same intermediary nodes (and neither has committed their changes to the
persisted data), neither will see any intermediary nodes and both will attempt to create them.

The fix is to have this logic create the intermediary nodes only if they don't exist; IOW, to ask that the
connector create them lazily. Unfortunately, an interface needed to be changed to add a new method (to create
nodes lazily), and several implementations changed to implement the new method. However, as they are new methods,
the impact should be isolated to this new logic. And since this logic is always used in all sequencing operations,
the impact should be pretty easily verified.

After I made the correction to the logic, I was not able to replicate the problem with the new test cases, even
after running almost a dozen times. I'll continue testing and running a full integration build before continuing.