MODE-1073 Correcting the logic to prevent creation of extra intermediary nodes
Turns out that even though I had altered the StreamSequencerAdapter to ask for the nodes to be created only if
they didn't exist, the GraphBatchDestination (the class with the implementation of the new method StreamSequencerAdapter
is now calling) was not using the correct graph methods given the desired NodeConflictBehavior. So the
GraphBatchDestination was always appending.

I'm correcting the GraphBatchDestination code to do the appropriate call for the supplied NodeConflictBehavior.