MODE-841 Added TextExtractor implementation that uses Apache Tika
Next added a 'modeshape-extractor-tika' module under the 'extensions/' directory, and set it up to use Tika 0.8.
(Hopefully Tika 0.9 will be out soon, as there are several good improvements. But it was originally scheduled for
a Dec 10, 2010 release but has obviously slipped.) The new Tika-based extractor finds Tika Parser implementations
via the ServiceLoader mechanism (0.9 has a much better implementation of this than 0.8), so our
'modeshape-extractor-tika' module is capable of using other Tika Parser implementations - simply by providing on
the classpath the JAR(s) containing the parsers and the "META-INF/services/org.apache.tika.parser.Parser" text files.

To enable the use of this extractor, the 'modeshape-extractor-tika-with-dependencies.jar' should be added to the
classpath (or the 'modeshape-extractor-tika' Maven artifact added to the application's POM), and the following
should be added to the ModeShape configuration file, immediately under the "configuration" root element:

  <mode:textExtractors>
    ... <!-- other extractors -->
    <mode:textExtractor jcr:name="Tika Text Extractors">
      <mode:description>Text extractors using Tika parsers</mode:description>
      <mode:classname>org.modeshape.extractor.tika.TikaTextExtractor</mode:classname>
      <!--
      A comma- or whitespace-delimited list of MIME types that are to be excluded. The following are excluded by default,
      but the default is completely overridden when this property is set. IOW, if you explicitly exclude any MIME types,
      be sure to list all of the MIME types you want to exclude. Exclusions always have a higher precedence than inclusions.
      -->
      <mode:excludedMimeTypes>application/x-archive,application/x-bzip,application/x-bzip2,application/x-cpio,
                              application/x-gtar,application/x-gzip,application/x-ta,application/zip,
                              application/vnd.teiid.vdb</mode:excludedMimeTypes>
      <!--
      A comma- or whitespace-delimited list of MIME types that are to be included. If this is used, then the extractor
      will include only those MIME types found in this list for which there is an available parser (unless the MIME type
      is also excluded). Including explicit MIME types is often easier if text is to be extracted for are only a few MIME types.
      -->
      <mode:includedMimeTypes>application/msword,application/vnd.oasis.opendocument.text</mode:includedMimeTypes>
    </mode:textExtractor>
    ... <!-- other extractors -->
  </mode:textExtractors>

Also added a 'org.modeshape.extractor.teiid.TeiidVdbTextExtractor' class that extracts the VDB name, description,
version, filename, and for each model the model's name, type, source name, source translator name, and source JNDI
name. At this point, no other information is extracted from VDBs, nor is there an extractor for Teiid model files.

To enable the VDB text extractor for the ModeShape engine, add the 'modeshape-sequencer-teiid-with-dependencies.jar'
to the classpath (or the 'modeshape-sequencer-teiid' Maven artifact to the application POM), and add the following
"mode:textExtractor" fragment under the "mode:textExtractors" element:

    <mode:textExtractor jcr:name="VDB Text Extractors">
      <mode:description>Extract text from Teiid VDB files</mode:description>
      <mode:classname>org.modeshape.extractor.teiid.TeiidVdbTextExtractor</mode:classname>
    </mode:textExtractor>

The commit is the 2nd commit for this feature.

This new feature is still not thoroughly tested, and so it is not ready to be merged into the official repository.