MODE-1083 Corrected the RepositoryNodeTypeManager's cache of property and child node definitions
Changed the ClusteredTest integration test to replicate the reported behavior.

In a clustered environment, whenever node types are registered (or re-registered) with the NodeTypeManager,
the RepositoryNodeTypeManager receiving the request persists the definitions in the system workspace,
and this generates events that are then received by all other RepositoryNodeTypeManager instances in the
cluster. They receive the events and update their in-memory representations of the node type definitions.

The RepositoryNodeTypeManager was correctly updating it's map of node type definitions keyed by the name,
but was not updating the cached maps of the property definitions and child node definitions with those
in the new/altered/removed node definitions. The fix is relatively minor, and involves updating these
maps in the same block of code that is updating the maps of node type definitions.

All unit and integration tests pass with these changes.