MODE-841 Added test cases for text extractors
Added several unit test cases for the Tika-based text extractor. There were some
strange behaviors with the text that Tika extracted: some text was duplicated
in the text extracted from the sample .pdf file; while the text extracted
from the .docx sample file was not the same as that from the .doc file.

Since no text extractors are added to the ModeShape configuration by default,
some of the configurations used for integration testing were changed to
use the Tika and Teiid VDB text extractors.