MODE-1085 Added use of Byteman bytecode injection for measuring timings of various methods.
Byteman 1.5.1 was recently introduced with better support for running within Maven and not needing an external installation of Byteman. This makes it much more attractive, and makes it possible to add integration tests that use Byteman to inject measurement code. One integration test was added with two test methods that use different (initially very simplistic) rule sets, verifying that we're able to get Byteman to work.