MODE-1065 Improvements to the way XML code is imported, reducing overall Request objects.
The XML importer functionality was changed to reduce the number of CreateNodeRequest and UpdatePropertiesRequest objects that are generated. Previously, a CreateNodeRequest usually had no properties, while there was frequently/usually an UpdatePropertiesRequest that set the properties. Now, with some minorly improved logic in the XmlHandler, importing XML typically results in a single CreateNodeRequest per XML element (though there are cases where not all properties can be determined before the first child node is discovered). In some tests, this results in a reduction of a little less than 50% in the number of requests produced from the XML import.

All unit and integration tests pass.