MODE-1085, MODE-1093, MODE-1065 Improved the thread-safety of the Lucene indexing components
Made a relatively small change introducing a read-write lock to limit the number of concurrent readers
and writers on a LuceneSearchEngine instance (e.g., a JcrRepository instance). With this change, only
one writer is allowed at a time per instance, or many concurrent readers are allowed per instance, but
no other combination is allowed.

The code was previously relying upon Lucene to do this (per the documentation), but from what I witnessed
during debug sessions there were definitely multiple IndexWriter implementations open and writing to
the same file system Directory instance.

With this change, all of the unit and integration tests pass.