MODE-1065 Improved performance of the GraphNamespaceRegistry and the ModeShapeEngine and JcrEngine startup
Several changes were made to the GraphNamespaceRegistry to reduce the number of requests sent to the connector.
Previously, each time a namespace was registered it re-read all the namespaces and issued at least one
update/create request. Upon startup of the JcrEngine, ten (or so) namespaces are registered, causing a large
number of extra read requests. Now, GraphNamespaceRegistry has a new 'register' method that takes a map
of namespace prefix-to-keys, and performs a single read of the existing namespaces, resulting in much
faster initialization of the namespace registry.

Additionally, the ModeShapeEngine and JcrEngine startup code was changed to read the configuration repository
content only once (or at most twice, if the configuration is adjusted). This results in faster startup
and reduced requests sent to the connector.

All unit and integration tests pass with these changes.