MODE-1092 Corrected the use of WEAKREFERENCE and REFERENCE values
Several minor problems were corrected so that it is possible to set a WEAKREFERENCE property on a node
and have that property remain a WEAKREFERENCE. Evidently there are no explicit checks for this in
the JCR 2.0 TCK, so this change adds a new unit test. Several of the errors were likely leftover
from when WEAKREFERENCE was added in JCR 2.0, but all corrections are relatively minor and reinforce
the similarity and conversion between REFERENCE and WEAKREFERENCE.

All unit and integration tests complete successfully with these changes.