MODE-1085 Improved the performance of the NamespaceRegistry implementations.
The GraphNamespaceRegistry was doing multiple reads of the entire namespaces subgraph whenever
a new namespace was registered. Thus, registering multiple namespaces (even a dozen) resulted
in a lot of extra reads. This change reduces the reads by providing a method to register
multiple namespaces, and to make the GraphNamespaceRegistry more efficient.

Note that these changes also use a new scheme in GraphNamespaceRegistry for storing
namespaces with generated prefixes. Prior to this, generated namespace prefixes were
stored as 'mode:namespace' nodes with names of "ns", and the SNS indexes were used
to define the number of the namespace prefix. This is acceptable for one namespace with a
generated prefix, but it can cause problems if there are multiple such namespaces
should 'ns001' be deleted (e.g., 'ns002' would be magically renamed to 'ns001' as the
name of the 'ns[2]' namespace node would be changed to 'ns[1]'). The new scheme
stores the full namespaces prefix as the name (e.g., 'ns001', 'ns002') and no longer
relies upon the SNS index. However, it is backward compatible, so if there are existing
content stores with the 'ns[1]' and 'ns[2]' style names, the GraphNamespaceRegistry
will work. (A workaround is to register an explicit prefix for any namespace
that has a generated prefix.)

All unit and integration tests pass with these changes.